+++ src/index.js
@@ -8,6 +8,8 @@ import renderCell from './lib/components/RenderCell.vue';
 import splitDownloadAndExport from './lib/components/SplitDownloadAndExport.vue';
 import renderSelectLoadmore from './lib/components/RenderSelectLoadmore.vue';
 import easyUpload from './lib/components/EasyUpload/index.vue';
+import ipageMask from "./lib/directives/mask/directive"
+import ipageMoney from "./lib/directives/money/index"
 export const config = $config
 export const IPage = ipage
 export const RenderCell = renderCell
@@ -18,6 +20,8 @@ export const ITableColumn = iTableColumn
 export const SplitDownloadAndExport = splitDownloadAndExport
 export const RenderSelectLoadmore = renderSelectLoadmore
 export const EasyUpload = easyUpload;
+export const IpageMask = ipageMask;
+export const IpageMoney = ipageMoney;
 
 const components = {IPage,RenderCell,ISearch,ITable,IForm,ITableColumn,SplitDownloadAndExport,RenderSelectLoadmore,EasyUpload};
 
@@ -26,6 +30,8 @@ const install = function (Vue) {
         const component = components[key];
         Vue.component(component.name||key, component);
     });
+    Vue.directive("ipage-mask", ipageMask);
+    Vue.directive("ipage-money", ipageMoney);
 };
 
 // 判断是否是直接引入文件，如果是，则自动安装组件